refactor: Remove unused code and add utility function for key retrieval
